Fix components with errors
Typescript needs some help for functions that return private classes
