+++ src/components/tooltip/index.js
@@ -1,1 +1,2 @@
-export default from './Tooltip';
+import Tooltip from './Tooltip';
+export default Tooltip;
